distributed all my functions from index.php to different files in functions folder. Renamed all user variables and functions to camel case
