+++ init.php
@@ -1,2 +1,6 @@
 <?php
-$config = require "config.php";
+$config = require_once("config.php");
+
+require_once("functions/database.php");
+require_once("functions/validation.php");
+require_once("functions/templates.php");
