--- tests/Unit/func.php
@@ -1,6 +1,0 @@
-<?php
-
-function makeUppercase(string $str): string
-{
-    return strtoupper($str);
-}
